Implement robust actor removing condition
Remove ports from actors that are arguments to delays as their
channels are removed. Then use port counts to determine if actor acts
purely as an arguments for delay actors. Previously checked if actor
had a single port, which led to bugs where the actor was connected to
more than one delay actor, or if it was also an argument to a
non-delay operator.